refactor(server): agent.routes.ts god function 분해 (#496)
createAgentRouter(865줄, complexity=164)를 5개 모듈로 분리:
- agent.routes.dto.ts: DTO 변환 함수 6개 (sessionDto, observationDto 등)
- agent.routes.utils.ts: 유틸 함수 11개 (writeError, requireString 등)
- agent.routes.events.ts: 이벤트 처리 (prepareEvent)
- agent.routes.injection.ts: 주입 빌더 (buildInjectionDetails 등)
- agent.routes.ts: buildRouterCtx + 20개 named handler + thin 라우터 팩토리

결과: agent.routes.ts 프로젝트 레벨 CRITICAL → 해소.
21개 spec 통과. 타입 오류 없음.

Closes #496. 상위: #495.
